+++ src/Common/Types.ts
@@ -1,6 +1,5 @@
 import { ReactNode, CSSProperties } from 'react'
 import { Placement } from 'tippy.js'
-import { VariableType } from './CIPipeline.Types'
 export interface ResponseType {
     code: number
     status: string
@@ -326,43 +325,4 @@ export interface CDMaterialResponseType {
     materials: any[]
     userApprovalConfig: UserApprovalConfigType
     requestedUserId: number
-}
-
-export enum ApplyPolicyToStage {
-    PRE_CI = 'PRE_CI',
-    POST_CI = 'POST_CI',
-    PRE_OR_POST_CI = 'PRE_OR_POST_CI',
-}
-
-export enum PluginRequiredStage {
-    PRE_CI = 'preBuildStage',
-    POST_CI = 'postBuildStage',
-    PRE_OR_POST_CI = 'PRE_OR_POST_CI',
-}
-
-export interface DefinitionSourceType {
-    projectName: string
-    isDueToProductionEnvironment: boolean
-    isDueToLinkedPipeline: boolean
-    policyName: string
-    appName?: string
-    clusterName?: string
-    environmentName?: string
-    branchNames?: string[]
-    ciPipelineName?: string
-}
-export interface MandatoryPluginDetailType {
-    id: number
-    name: string
-    description?: string
-    requiredIn: PluginRequiredStage
-    applied?: boolean
-    inputVariables?: VariableType[]
-    outputVariables?: VariableType[]
-    definitionSources?: DefinitionSourceType[]
-}
-export interface MandatoryPluginDataType {
-    pluginData: MandatoryPluginDetailType[]
-    isValidPre: boolean
-    isValidPost: boolean
 }
